(validate_path): Cast strlen value to `unsigned long' so it matches `%ld' format even on 32-bit systems.

+++ src/pathchk.c
@@ -374,7 +374,7 @@ validate_path (char *path, int portability)
   if (strlen (path) > (size_t) path_max)
     {
       error (0, 0, _("path `%s' has length %ld; exceeds limit of %ld"),
-	     path, strlen (path), path_max);
+	     path, (unsigned long) strlen (path), path_max);
       return 1;
     }
